Lab 11 done.

+++ user/dumbfork.c
@@ -49,7 +49,7 @@ dumbfork(void) {
   // will return 0 instead of the envid of the child.
   envid = sys_exofork();
   if (envid < 0)
-    panic("sys_exofork: %i", envid);
+    panic("sys_exofork: %d", (int)envid);
   if (envid == 0) {
     // We're the child.
     // The copied value of the global variable 'thisenv'
